% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoR_trf.R
\name{geoR_trf}
\alias{geoR_trf}
\title{Default transition function in geoR}
\usage{
geoR_trf(x)
}
\arguments{
\item{x}{numeric vector of length 2 with cost/conductivity non-negative
  numbers.}
}
\value{
cost/conductivity value of the transition between neighbouring cells.
}
\description{
Default function to use with \code{\link[gdistance]{transition}}. Defined as
the geometric mean.
}
\details{
To use another function, override this one by defining it using the same
name.
}
\examples{
geoR_trf(c(0, 1))

   # redefine function
   geoR_trf <- mean

   geoR_trf(c(0, 1))
}

