% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distmatGen.R
\name{idx_isol}
\alias{idx_isol}
\title{Identify isolated locations}
\usage{
idx_isol(x)
}
\description{
Defined as points whith cost distances to every other
point of either 0 or Infinite.
}
\details{
param x distance matrix.
}

