% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loglik.GRF}
\alias{loglik.GRF}
\title{Log-likelihood}
\usage{
loglik.GRF(geodata, coords = geodata$coords, data = geodata$data,
  obj.model = NULL, cov.model = "exp", cov.pars, nugget = 0,
  kappa = 0.5, lambda = 1, psiR = 1, psiA = 0, trend = "cte",
  method.lik = "ML", compute.dists = TRUE, realisations = NULL)
}
\description{
Log-likelihood
}

