% Generated by roxygen2 (4.0.2): do not edit by hand
\name{varcov.spatial}
\alias{varcov.spatial}
\title{Variance-covariance}
\usage{
varcov.spatial(coords = NULL, dists.lowertri = NULL, cov.model = "matern",
  kappa = 0.5, nugget = 0, cov.pars = stop("no cov.pars argument"),
  inv = FALSE, det = FALSE, func.inv = c("cholesky", "eigen", "svd",
  "solve"), scaled = FALSE, only.decomposition = FALSE, sqrt.inv = FALSE,
  try.another.decomposition = TRUE, only.inv.lower.diag = FALSE, ...)
}
\description{
Variance-covariance
}

