% Generated by roxygen2 (4.0.2): do not edit by hand
\name{variog}
\alias{variog}
\title{cost-based empirical variogram}
\usage{
variog(geodata, coords = geodata$coords, data = geodata$data,
  uvec = "default", breaks = "default", trend = "cte", lambda = 1,
  option = c("bin", "cloud", "smooth"), estimator.type = c("classical",
  "modulus"), nugget.tolerance, max.dist, pairs.min = 2, bin.cloud = FALSE,
  direction = "omnidirectional", tolerance = pi/8,
  unit.angle = c("radians", "degrees"), angles = FALSE, dists.mat, messages,
  ...)
}
\description{
All the arguments work as in \code{\link[geoR]{variog}}, except the
additional argument \code{dd.dists.mat}, which takes a symmetric matrix of
distances between observation locations
}
\examples{
## geodata structure with transformed covariates
   data(noise)
   covarnames=sapply(1:3, function(x) paste("d2TV", x, sep=""))
   obs.df <- data.frame(Leq=obs$Leq,
                         1/(1+(as.data.frame(obs)[covarnames]/20)^2))
   obs.gd <- as.geodata(cbind(coordinates(obs), obs.df),
                        data.col="Leq",
                        covar.col=c('d2TV1','d2TV2','d2TV3'))

   ## compute euclidean and cost-based empirical variograms
   vg.std <- variog(obs.gd, trend=~d2TV1*(d2TV2+d2TV3))
   vg.dmat <- variog(obs.gd, trend=~d2TV1*(d2TV2+d2TV3), dists.mat=dd.distmat)

   ## plot and compare empirical variograms
   plot(vg.std, type = 'l', col = 'darkred',
        main = 'Euclidean (red) vs. cost-based (gray) empirical variograms')
   lines(vg.dmat, type = 'l', col = 'darkgray')
}

