% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dists.R
\name{vecdist}
\alias{vecdist}
\title{Vector of reciprocal distances}
\usage{
vecdist(x)
}
\arguments{
\item{x}{a \code{data.frame} with planar coordinates}
}
\description{
If there is a custom definicion of distances, it uses it. Otherwise, it
computes the reciprocal distances of points in x.
}
\details{
If there exists a matrix object named
\code{.personal.definition.of.distances} in the global environment, the
function returns the lower triangular part of it as a vector, regardless of
\code{x}. Otherwise, the Euclidean distances between points in \code{x} are
returned as a vector.
}

