% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distmatGen.R
\name{distmatGen}
\alias{distmatGen}
\title{Compute cost-based distances}
\usage{
distmatGen(pts, costsurf, ret, directions = 16)
}
\arguments{
\item{pts}{A SpatialPointsDataFrame with a defined projection}

\item{costsurf}{cost surface Raster}

\item{ret}{specify whether to return distances between obs-obs ("o"), obs-loc
("l"), or both ("b")}

\item{directions.}{See \code{\link[gdistance]{transition}}.}
}
\description{
Generate a cost-based distance matrix among observations and/or between
observations and prediction locations.
}
\details{
Use the centroids of the cost surface raster cells as prediction locations.

If a cost surface raster is not supplied a uniform (cost 1) grid will be
generated to match to bounding box of pts
}
\examples{
data(noise)
 r <- raster(nrows=40,ncols=40,resolution=1)
 r <- setExtent(r,extent(obs),keepres=T)
 aggfac <- 5
 costsurf <- rasterize(malilla,r,background=aggfac,field=rep(10000,length(polygons(malilla))))
 costsurf <- reclassify(costsurf,c(aggfac+1,Inf,NA))
 costsurf <- aggregate(costsurf,aggfac,expand=T,fun=max,na.rm=T)

 res <- distmatGen(obs,costsurf,ret="o")
 plot(dd.distmat[,51],res[,51],ylim=c(0,600),xlim=c(0,600))
}

