% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distmatGen.R
\name{distmatGen}
\alias{distmatGen}
\title{Generate a cost-based distance matrix among SpatialPoints}
\usage{
distmatGen(pts, costsurf, ret)
}
\arguments{
\item{pts}{A SpatialPointsDataFrame with a defined projection}

\item{costsurf}{cost surface Raster}

\item{ret}{specify whether to return distances between obs-obs ("o"), obs-loc ("l"), or both ("b")}
}
\description{
Generate a cost-based distance matrix among SpatialPoints
}
\examples{
data(noise)
r<-raster(nrows=40,ncols=40,resolution=1)
r<-setExtent(r,extent(obs),keepres=T)
aggfac<-2
costsurf<-rasterize(malilla,r,background=aggfac,field=rep(10000,length(polygons(malilla))))
costsurf<-reclassify(costsurf,c(aggfac+1,Inf,NA))
costsurf<-aggregate(costsurf,aggfac,expand=T,fun=max,na.rm=T)

res<-distmatGen(obs,costsurf,ret="o")
plot(dd.distmat[,51],res[,51],ylim=c(0,600),xlim=c(0,600))
}

