% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likGRF.R
\name{loglik.GRF}
\alias{loglik.GRF}
\title{Log-likelihood}
\usage{
loglik.GRF(geodata, coords = geodata$coords, data = geodata$data,
  obj.model = NULL, cov.model = "exp", cov.pars, nugget = 0,
  kappa = 0.5, lambda = 1, psiR = 1, psiA = 0, trend = "cte",
  method.lik = "ML", compute.dists = TRUE, realisations = NULL)
}
\arguments{
\item{geodata}{a list containing elements \code{coords} and
    \code{data} as described next. 
    Typically an object of the class \code{"geodata"} - a \pkg{geoR}
    data-set.
    If not provided the arguments
    \code{coords} and \code{data} must be provided instead.  }

\item{coords}{an \eqn{n \times 2}{n x 2} matrix, each row containing Euclidean
    coordinates of the \emph{n} data locations. By default it takes the
    element \code{coords} of the argument \code{geodata}.  }

\item{data}{a vector with data values. By default it takes the
    element \code{data} of the argument \code{geodata}.  }

\item{obj.model}{a object of the class \code{variomodel} with a fitted
    model. Tipically an output of 
    \code{\link{likfit}} or \code{\link{variofit}}. }

\item{cov.model}{a string specifying the model for the correlation
    function. For further details see 
    documentation for \code{\link{cov.spatial}}. }

\item{cov.pars}{a vector with 2 elements with values of  the covariance parameters
    \eqn{\sigma^2}{sigma^2} (partial sill) and \eqn{\phi}{phi} (range
    parameter).  }

\item{nugget}{value of the nugget parameter. Defaults to \eqn{0}.  }

\item{kappa}{value of the smoothness parameter. Defaults to
    \eqn{0.5}.  }

\item{lambda}{value of the Box-Cox tranformation parameter. Defaults
    to \eqn{1}.  }

\item{psiR}{value of the anisotropy ratio parameter. Defaults to
    \eqn{1}, corresponding to isotropy.  }

\item{psiA}{value (in radians) of the anisotropy rotation parameter. Defaults to zero.}

\item{trend}{specifies the mean part of the model.
    The options are:
    \code{"cte"} (constant mean),
    \code{"1st"} (a first order polynomial
    on the coordinates), \code{"2nd"} (a second order polynomial
    on the coordinates), or a formula of the type \code{~X} where \code{X}
    is a matrix with the covariates (external trend). Defaults to \code{"cte"}.  }

\item{method.lik}{options are \code{"ML"} for likelihood and \code{"REML"} for
    restricted likelihood. Defaults to \code{"ML"}.  }

\item{compute.dists}{for internal use with other function. Don't
    change the default unless you know what you are doing.  }

\item{realisations}{optional. A vector indicating replication number
    for each data. For more details see \code{\link{as.geodata}}. }
}
\description{
Same as geoR's \code{\link[geoR]{loglik.GRF}} but taking into account non-Euclidean
distances if pertinent.
}
\seealso{
\code{\link[geoR]{loglik.GRF}}
}

